Address PR review round 2: overflow, ordering, abort-on-unknown-form
- DAP::InitLocal: bound sun_path against sizeof(sun_path), reject overflow
  (strcpy -> strncpy with explicit NUL).
- RealtimeWatchSampler::AddSubscription: reject count==0 or address+count
  overflow (new kInvalidWatchId=0 sentinel) so Tick's address+i is safe.
- DapSession: defer entry-stop until BOTH launch/attach AND configurationDone
  have arrived via MaybeFireEntryStop(); a configDone-before-launch no longer
  drops a later launch's stopOnEntry override. Remember launch kind for the
  event reason. New configDone-before-launch stopOnEntry=false regression
  test.
- DapDebugController::GetSource: drop unreachable current_line==0 check;
  empty content already encodes the no-lines condition.
- PPCSymbolDB::HasDenseLineInfoInRange: compute end in u64 so a range
  spanning the top of the 32-bit address space doesn't wrap to 0.
- DwarfReader::ParseDie: unknown attribute form no longer aborts the entire
  CU walk; log a warning and return true so the caller advances via
  length/sibling (renamed regression test to
  ParseToleratesUnknownAttributeFormAndContinues).
- Core.cpp: only mark debugger_enabled after DAP::IsActive() confirms the
  listener bound, so a bad socket/port config no longer hangs the core
  paused with no DAP server to resume it.

229 DAP/DWARF/test suite entries pass.

+++ Source/Core/Core/PowerPC/PPCSymbolDB.cpp
@@ -193,11 +193,14 @@ bool PPCSymbolDB::HasDenseLineInfoInRange(const u32 start, const u32 size) const
   if (m_line_table.empty() || size == 0)
     return false;
 
-  const u32 end = start + size;
+  // DESNOTE(jbarber, 2026-07-21): Compute the exclusive end in u64 so a range
+  // spanning the very top of the 32-bit address space doesn't wrap to 0 and
+  // erroneously match low-memory line entries.
+  const u64 end = static_cast<u64>(start) + static_cast<u64>(size);
   std::set<u32> distinct_lines;
   for (const auto& [address, entry] : m_line_table)
   {
-    if (address < start || address >= end)
+    if (static_cast<u64>(address) < start || static_cast<u64>(address) >= end)
       continue;
 
     distinct_lines.insert(entry.line);
